redirect stdout/stderr to improve secret
